Add code to call api

+++ src/main/java/moneytransferapp/MoneyTransferWorkflow.java
@@ -9,6 +9,8 @@ public interface MoneyTransferWorkflow {
 
     // The Workflow method is called by the initiator either via code or CLI.
     @WorkflowMethod
-    void transfer(String fromAccountId, String toAccountId, String referenceId, double amount);
+    void transfer(String fromAccountId, String toAccountId, String referenceId, int amount);
+
+
 }
 // @@@SNIPEND
